requirements.txt
# Core AI/ML dependencies
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Blockchain dependencies
web3>=6.0.0
eth-account>=0.8.0
py-solc-x>=1.12.0

# Web framework (optional)
fastapi>=0.85.0
uvicorn>=0.18.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.9.0
aiofiles>=0.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0

# Development
black>=22.0.0
flake8>=4.0.0